Fix dev branch

+++ src/main/java/ua/shpp/eqbot/service/UserService.java
@@ -2,7 +2,6 @@ package ua.shpp.eqbot.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
@@ -18,20 +17,16 @@ import ua.shpp.eqbot.validation.UserValidateService;
 public class UserService {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserService.class);
     private final UserRepository userRepository;
-    private final String dtoCacheName = "cacheUserDto";
-    private final CacheManager cacheManager;
-    private final UserValidateService userValidateService;
+    private static final String DTO_CACHE_NAME = "cacheUserDto";
 
 
-    public UserService(UserRepository userRepository, UserValidateService userValidateService, CacheManager cacheManager) {
+    public UserService(UserRepository userRepository, UserValidateService userValidateService) {
         this.userRepository = userRepository;
-        this.userValidateService = userValidateService;
-        this.cacheManager = cacheManager;
     }
 
     @Cacheable(cacheNames = "cacheEntity", key = "#telegramId")
     public UserEntity getEntity(Long telegramId) {
-        LOGGER.info("get userEntity by telergamId {}", telegramId);
+        LOGGER.info("get userEntity by telegramId {}", telegramId);
         UserDto dto = getDto(telegramId);
         return dto != null ? UserMapper.INSTANCE.userDTOToUserEntity(dto) : userRepository.findByTelegramId(telegramId);
     }
@@ -43,25 +38,24 @@ public class UserService {
         return userRepository.save(userEntity);
     }
 
-    @CachePut(cacheNames = dtoCacheName, key = "#userDto.telegramId")
+    @CachePut(cacheNames = DTO_CACHE_NAME, key = "#userDto.telegramId")
     public UserDto saveDto(UserDto userDto) {
         LOGGER.info("save userDto {}", userDto);
         return userDto;//****
     }
 
-    @Cacheable(cacheNames = dtoCacheName, key = "#telegramId")
+    @Cacheable(cacheNames = DTO_CACHE_NAME, key = "#telegramId")
     public UserDto getDto(Long telegramId) {
         LOGGER.info("get userDto by telegramId {}", telegramId);
         return null;
     }
 
     @Transactional
-    @CacheEvict(cacheNames = {dtoCacheName, "cacheEntity"}, key = "#telegramId")
+    @CacheEvict(cacheNames = {DTO_CACHE_NAME, "cacheEntity"}, key = "#telegramId")
     public boolean remove(Long telegramId) {
         LOGGER.info("delete userDto and All entity");
         UserEntity entity = getEntity(telegramId);
-        if (entity != null)
-            userRepository.delete(getEntity(telegramId));
+        if (entity != null) userRepository.delete(getEntity(telegramId));
         return true;
     }
 }
